% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konstruktory.R
\name{konstruktor_okregow}
\alias{konstruktor_okregow}
\title{Konstruktor macierzy okregów wyborczych}
\usage{
konstruktor_okregow(nazwa)
}
\arguments{
\item{nazwa}{scieżka do pliku CSV lub XLS pobranego ze strony PKW}
}
\value{
macierz o nazwie 'okregi'
}
\description{
Funkcja tworzy macierz, w której znajduja sie informacje dotyczace
okregów wyborczych - liczby ludnosci oraz dostepnych mandatów.
}
\examples{
konstruktor_okregow("okregi_sejm.csv")
}
