% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konstruktor_wynikow.R
\name{konstruktor_wynikow}
\alias{konstruktor_wynikow}
\title{Konstruktor macierzy wyników wyborów}
\usage{
konstruktor_wynikow(nazwa, ...)
}
\arguments{
\item{nazwa}{sciezka do pliku CSV lub XLS pobranego ze strony PKW}

\item{...}{numery kolumn w ktorych znajduja sie wyniki interesujacych nas komitetow}
}
\value{
macierz klasy 'macierz_wynikow'
}
\description{
Funkcja tworzy macierz, w ktorej każda kolumna zawiera wyniki wybranych
komitetów wyborczych. Każdy wiersz to inny okręg. Stworzona macierz może zostać później
użyta w funkcji wybory_rok w celu symulacji podziału mandatów.
}
\examples{
konstruktor_wynikow("sejm_wyniki_2019.xlsx", 9, 11, 12, 14, 16)
}
