% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_wyniki.R
\name{pobierz_wyniki}
\alias{pobierz_wyniki}
\title{Funkcja pobierająca dane dot. wyników wyborów do Sejmu RP}
\usage{
pobierz_wyniki(rok)
}
\arguments{
\item{rok}{rok, w którym odbyły się wybory parlamentarne (do wyboru: 2007, 2011, 2015, 2019)}
}
\value{
pobrane pliki w folderze 'dane_wybory'
}
\description{
Funkcja pobiera pliki .xlsx i .csv z wynikami wyborów
dla podanego przez użytkownika roku, a także z danymi o okręgach wyborczych
}
\examples{
pobierz_wyniki(2019)
}
