% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wybory_app.R
\name{wybory_app}
\alias{wybory_app}
\title{Aplikacja Shiny}
\usage{
wybory_app()
}
\arguments{
\item{rok}{rok, w którym odbyły się wybory parlamentarne (do wyboru: 2007, 2011, 2015, 2019)}
}
\value{
aplikacja shiny
}
\description{
Funkcja pobiera pliki .xlsx i .csv z wynikami wyborów
dla podanego przez użytkownika roku, a także z danymi o okręgach wyborczych
}
\examples{
pobierz_wyniki(2019)
}
