% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wybory_pl.R
\name{wybory_pl}
\alias{wybory_pl}
\title{Obliczanie rozkladu mandatow przy rownomiernym poparciu w skali kraju}
\usage{
wybory_pl(..., frekwencja = 100)
}
\arguments{
\item{...}{wyniki wyborcze komitetow w \% (max. 10 komitetów)}

\item{frekwencja}{frekwencja wyborcza, domyslnie ustawiona na 100\%}
}
\value{
macierz z wynikami oraz wykres obrazujacy wyniki
}
\description{
Funkcja tworzy macierz zawierajacą informacje o liczbie mandatow poselskich przyznanych
komitetom wg trzech roznych metod w przypadku gdyby wszystkie komitety otrzymaly rownomierne poparcie w skali kraju.
Wyswietlany także jest wykres zawierajacy wspomniane informacje,
a takze wizualizujacy realne poparcie w okregu w odniesieniu do dostepnych mandatow w skali kraju.
}
\examples{
wybory_pl(30, 29, 10, 7, 6)
}
