% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wybory_rok.R
\name{wybory_rok}
\alias{wybory_rok}
\title{Obliczanie rozkladu mandatow dla wyników z danego roku wg różnych metod}
\usage{
wybory_rok(wyniki)
}
\arguments{
\item{wyniki}{obiekt klasy 'macierz_wynikow' tworzony przez funkcję 'konstruktor_wynikow'}
}
\value{
macierz z wynikami oraz wykres obrazujacy wyniki
}
\description{
Funkcja tworzy macierz zawierajacą informacje o liczbie mandatow poselskich przyznanych
komitetom wg trzech roznych metod w wyborach parlamentarnych wybranego roku.
Wyswietlany także jest wykres zawierajacy wspomniane informacje,
a takze wizualizujacy realne poparcie w okregu w odniesieniu do dostepnych mandatow w skali kraju.
Do poprawnego działania funkcji wymagane jest pobranie arkusza z wynikami wyborów z PKW oraz
przetworzenie ich przy pomocy funkcji 'konstruktor_wynikow'.
}
\examples{
konstruktor_wynikow("sejm_wyniki_2019.xlsx", 9, 11, 12, 14, 16)
wybory_rok(wybory_2019)
}
