% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_mandaty.R
\name{wykres_mandaty}
\alias{wykres_mandaty}
\title{Wykres wizualizujacy wyniki wyborow w zaleznosci od wybranej metody podzialu mandatow}
\usage{
wykres_mandaty(wyniki)
}
\arguments{
\item{wyniki}{obiekt z wynikami podziału mandatów powstaly w wyniku funkcji wybory_rok}
}
\value{
wykres
}
\description{
Funkcja zwraca wykres przedstawiajacy rozklad mandatow na podstawie wynikow wyborow w zaleznosci od wybranej metody podzialu mandatow.
Kolorowe slupki przedstawiaja liczbe uzyskanych mandatow, a czarne slupki bez wypelnienia obrazuja ile mandatow przyslugiwaloby komitetowi gdyby zastosowac czysto proporcjonalny podzial mandatow.
}
\examples{
wykres_mandaty(wyniki_2019)
}
