% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres.R
\name{wykres_wyniki}
\alias{wykres_wyniki}
\title{Wykres eksploracyjny - wyniki wyborow}
\usage{
wykres_wyniki(nazwa, kol1, kol2, kol3, kol4, kol5)
}
\arguments{
\item{nazwa}{sciezka do pliku z wynikami wyborow pobranego ze strony PKW}

\item{kol1, kol2, kol3, kol4, kol5}{numery kolumn w ktorych znajduja sie wyniki interesujacych nas pieciu komitetow}
}
\value{
wykres pudelkowy obrazujacy rozklad wynikow partii w okregach
}
\description{
Funkcja ta tworzy wykres pudelkowy obrazujacy rozklad wynikow poszczegolnych komitetow wyborczych w okregach
}
\examples{
wykres_wyniki("sejm_wyniki_2019.xlsx", kol1 = 9, kol2 = 11, kol3 = 12, kol4 = 14, kol5 = 16)
}
