% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konstruktor_okregow.R
\name{konstruktor_okregow}
\alias{konstruktor_okregow}
\title{Konstruktor ramki danych z informacjami o okregach wyborczych}
\usage{
konstruktor_okregow(nazwa)
}
\arguments{
\item{nazwa}{sciezka do pliku CSV lub XLS pobranego ze strony PKW}
}
\value{
ramka danych o nazwie 'okregi'
}
\description{
Funkcja tworzy ramke danych, w ktorej znajduja sie informacje dotyczace
okregow wyborczych - liczby ludnosci oraz dostepnych mandatow.
}
\examples{
konstruktor_okregow("dane_wybory/okregi2019.xlsx")
}
