% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konstruktor_okregow.R
\name{konstruktor_okregow}
\alias{konstruktor_okregow}
\title{Konstruktor macierzy okregow wyborczych}
\usage{
konstruktor_okregow(nazwa)
}
\arguments{
\item{nazwa}{sciezka do pliku CSV lub XLS pobranego ze strony PKW}
}
\value{
macierz o nazwie 'okregi'
}
\description{
Funkcja tworzy macierz, w ktorej znajduja sie informacje dotyczace
okregow wyborczych - liczby ludnosci oraz dostepnych mandatow.
}
\examples{
konstruktor_okregow("okregi2019.csv")
}
