% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_wyniki.R
\name{pobierz_wyniki}
\alias{pobierz_wyniki}
\title{Funkcja pobierająca dane dot. wyników wyborów do Sejmu RP}
\usage{
pobierz_wyniki(rok, path = getwd())
}
\arguments{
\item{rok}{rok, w którym odbyły się wybory parlamentarne (do wyboru: 2007, 2011, 2015, 2019)}

\item{path}{folder, w którym mają zostać zapisane pobrane pliki, domyślnie jest to aktualny folder roboczy}
}
\value{
pobrane pliki w folderze 'dane_wybory'
}
\description{
Funkcja pobiera pliki .xlsx i .csv z wynikami wyborów
dla podanego przez użytkownika roku, a także z danymi o okręgach wyborczych
}
\examples{
pobierz_wyniki(2019)
}
