% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wybory_okreg.R
\name{wybory_okreg}
\alias{wybory_okreg}
\title{Obliczanie rozkladu mandatow w okregu}
\usage{
wybory_okreg(..., okreg, frekwencja = 100)
}
\arguments{
\item{...}{wyniki wyborcze komitetow w \% (max. 10 komitetów)}

\item{okreg}{liczba z przedzialu (1, 41); numer okregu, w ktorym chcemy obliczyc rozklad mandatow}

\item{frekwencja}{frekwencja wyborcza, domyslnie ustawiona na 100\%}
}
\value{
ramka danych z wynikami
}
\description{
Funkcja tworzy ramke danych zawierajaca informacje o liczbie mandatow z okregu przyznanych
komitetom wg trzech roznych metod.
}
\examples{
wybory_okreg(30, 29, 10, 7, 6, okreg = 4)
}
