% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wybory_rok.R
\name{wybory_rok}
\alias{wybory_rok}
\title{Obliczanie rozkladu mandatow dla wynikow z danego roku wg roznych metod}
\usage{
wybory_rok(wyniki)
}
\arguments{
\item{wyniki}{ramka danych z wynikami stworzona przez funkcję 'konstruktor_wynikow'}
}
\value{
ramka danych z wynikami
}
\description{
Funkcja tworzy macierz zawierajaca informacje o liczbie mandatow poselskich przyznanych
komitetom wg trzech roznych metod w wyborach parlamentarnych wybranego roku.
Do poprawnego dzialania funkcji wymagane jest pobranie arkusza z wynikami wyborow z PKW oraz
przetworzenie ich przy pomocy funkcji 'konstruktor_wynikow'.
}
\examples{
konstruktor_wynikow("sejm_wyniki_2019.xlsx", 9, 11, 12, 14, 16)
wybory_rok(wybory_2019)
}
