% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_wyniki.R
\name{wykres_wyniki}
\alias{wykres_wyniki}
\title{Wykres eksploracyjny - wyniki wyborow}
\usage{
wykres_wyniki(wyniki, type = "boxplot")
}
\arguments{
\item{wyniki}{ramka danych z wynikami wyborow stworzona przy pomocy funkcji konstruktor_wynikow}

\item{type}{typ wykresu, jaki chcemy stworzyc
dostepne typy: boxplot, violin, dotplot, scatter}
}
\value{
wykres pudelkowy obrazujacy rozklad wynikow partii w okregach
}
\description{
Funkcja ta tworzy wykres pudelkowy obrazujacy rozklad wynikow poszczegolnych komitetow wyborczych w okregach
}
\examples{
wykres_wyniki(wybory_2019, type = "boxplot")
}
