% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres.R
\name{wykres_wyniki}
\alias{wykres_wyniki}
\title{Wykres eksploracyjny - wyniki wyborow}
\usage{
wykres_wyniki(nazwa, ...)
}
\arguments{
\item{nazwa}{sciezka do pliku z wynikami wyborow pobranego ze strony PKW}

\item{...}{numery kolumn w ktorych znajduja sie wyniki interesujacych nas komitetow}
}
\value{
wykres pudelkowy obrazujacy rozklad wynikow partii w okregach
}
\description{
Funkcja ta tworzy wykres pudelkowy obrazujacy rozklad wynikow poszczegolnych komitetow wyborczych w okregach
}
\examples{
wykres_wyniki("sejm_wyniki_2019.xlsx", 9, 11, 12, 14, 16)
}
